detamvc/templates/core/requirements.txt
fastapi
jinja2
deta
detamvc
python-multipart
python-dotenv